requirements.txt
streamlit
scikit-learn
matplotlib
seaborn
joblib
numpy
pillow
